autogen-thematic-etf-multiagent/requirements.txt
# Core AG2 (AutoGen) and AI dependencies
# AG2 is the continuation of AutoGen with better multi-provider support
ag2>=0.8.0
# Alternative: pyautogen (alias for ag2)
# Install with extras for different providers
# pip install 'ag2[anthropic,openai,gemini]'

# LLM Provider SDKs (installed as dependencies by ag2 extras, but listed for clarity)
openai>=1.12.0
google-generativeai>=0.3.0
anthropic>=0.18.0

# Web Interface
gradio>=4.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.5.0

# Optional: Data handling for ETF analysis
pandas>=2.0.0
numpy>=1.24.0

# Development tools (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
